+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Increment `react-redux` to `v7`, `redux-form` to `v8`, `@folio/stripes` to `v6`. Refs UID-38.
 * Increment `@folio/stripes-cli` to `v2`. Refs UID-47.
+* Use new SSC `ConfigReduxForm` export in place of `ConfigForm`. Refs UID-72.
 
 ## [4.0.0](https://github.com/folio-org/ui-developer/tree/v4.0.0) (2020-10-09)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v3.0.0...v4.0.0)